Update README.md
So I see when I create a new branch the commit button now reads : Propose changes.  That makes sense.  I also just followed the link below that says "Learn more about pull requests" and there's a lot of info there so I'll bookmark it and circle back to it later.